+++ config/silverstripe-6-0.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use Netwerkstatt\SilverstripeRector\Rector\Control\ReplaceHasCurrWithCurrRector;
 use Netwerkstatt\SilverstripeRector\Rector\DataObject\DataObjectGetByIdToByIDRector;
+use Netwerkstatt\SilverstripeRector\Rector\Misc\RemoveSilverstripeDeprecationCommentRector;
 use Rector\Config\RectorConfig;
 use Rector\Renaming\Rector\ClassConstFetch\RenameClassConstFetchRector;
 use Rector\Renaming\Rector\MethodCall\RenameMethodRector;
@@ -129,9 +130,11 @@ return static function (RectorConfig $rectorConfig): void {
         new MethodCallRename('SilverStripe\CMS\Controllers\LeftAndMainRecordIconsExtension', 'generatePageIconsCss',
             'generateRecordIconsCss'),
         new MethodCallRename('SilverStripe\Forms\Form', 'validationResult', 'validate'),
+        new MethodCallRename('SilverStripe\Forms\FormField', 'Value', 'getFormattedValue'),
         new MethodCallRename('SilverStripe\Forms\TextareaField', 'ValueEntities', 'getFormattedValueEntities'),
         new MethodCallRename('SilverStripe\Model\List\ListDecorator', 'TotalItems', 'getTotalItems'),
         new MethodCallRename('SilverStripe\Model\List\PaginatedList', 'TotalItems', 'getTotalItems'),
+        new MethodCallRename('SilverStripe\Control\Director', 'get_session_environment_type', 'get_environment_type'),
 
         // Member hooks
         new MethodCallRename('SilverStripe\Core\Extension', 'afterMemberLoggedIn', 'onAfterMemberLoggedIn'),
@@ -161,6 +164,28 @@ return static function (RectorConfig $rectorConfig): void {
     ]);
     $rectorConfig->rule(ReplaceHasCurrWithCurrRector::class);
     $rectorConfig->rule(DataObjectGetByIdToByIDRector::class);
+    $rectorConfig->ruleWithConfiguration(RemoveSilverstripeDeprecationCommentRector::class, [
+        'SilverStripe\Model\List\ListDecorator::getTotalItems' => [
+            'message' => 'ListDecorator::TotalItems() has been deprecated. Use getTotalItems() instead.',
+            'link' => 'https://docs.silverstripe.org/en/5/changelogs/5.4.0/#deprecated-api',
+        ],
+        'SilverStripe\Model\List\PaginatedList::getTotalItems' => [
+            'message' => 'PaginatedList::TotalItems() has been deprecated. Use getTotalItems() instead.',
+            'link' => 'https://docs.silverstripe.org/en/5/changelogs/5.4.0/#deprecated-api',
+        ],
+        'SilverStripe\Forms\TextareaField::getFormattedValueEntities' => [
+            'message' => 'TextareaField::ValueEntities() has been deprecated. Use getFormattedValueEntities() instead.',
+            'link' => 'https://docs.silverstripe.org/en/5/changelogs/5.4.0/#deprecated-api',
+        ],
+        'SilverStripe\Forms\FormField::getFormattedValue' => [
+            'message' => 'FormField::Value() has been deprecated. It will be replaced by getFormattedValue() and getValue().',
+            'link' => 'https://docs.silverstripe.org/en/5/changelogs/5.4.0/#deprecated-api',
+        ],
+        'SilverStripe\Control\Director::get_environment_type' => [
+            'message' => 'Director::get_session_environment_type() has been deprecated. Use Director::get_environment_type() instead.',
+            'link' => 'https://docs.silverstripe.org/en/5/changelogs/5.4.0/#deprecated-api',
+        ],
+    ]);
     $rectorConfig->ruleWithConfiguration(RenameClassConstFetchRector::class, [
         new RenameClassAndConstFetch('SilverStripe\Admin\LeftAndMain', 'SCHEMA_HEADER',
             'SilverStripe\Forms\Schema\FormSchema', 'SCHEMA_HEADER'),