+++ stubs/SilverStripe/Control/Director.php
@@ -5,4 +5,6 @@ namespace SilverStripe\Control;
 class Director
 {
     public static $alternate_public_dir;
+    public static function get_session_environment_type() {}
+    public static function get_environment_type() {}
 }