+++ stubs/SilverStripe/Forms/FormField.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SilverStripe\Forms;
+
+class FormField
+{
+    public function Value() {}
+    public function getFormattedValue() {}
+}