+++ stubs/SilverStripe/ORM/FieldType/DBEnum.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SilverStripe\ORM\FieldType;
+
+if (class_exists('SilverStripe\ORM\FieldType\DBEnum')) {
+    return;
+}
+
+class DBEnum
+{
+    public function reset()
+    {
+    }
+}