+++ tests/Misc/RemoveSilverstripeDeprecationCommentRector/RemoveSilverstripeDeprecationCommentRectorTest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Netwerkstatt\SilverstripeRector\Tests\Misc\RemoveSilverstripeDeprecationCommentRector;
+
+use Rector\Testing\PHPUnit\AbstractRectorTestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
+
+final class RemoveSilverstripeDeprecationCommentRectorTest extends AbstractRectorTestCase
+{
+    #[DataProvider('provideData')]
+    public function test(string $filePath): void
+    {
+        $this->doTestFile($filePath);
+    }
+
+    public static function provideData(): \Iterator
+    {
+        return self::yieldFilesFromDirectory(__DIR__ . '/Fixture');
+    }
+
+    public function provideConfigFilePath(): string
+    {
+        return __DIR__ . '/config/configured_rule.php';
+    }
+}