+++ tests/Set/Silverstripe54To60/config/configured_rule.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Netwerkstatt\SilverstripeRector\Set\SilverstripeSetList;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->sets([
+        SilverstripeSetList::SS_5_4,
+        SilverstripeSetList::SS_6_0,
+    ]);
+};